display list of temporary SampleData with RecyclerView
Created a hard-coded list of SampleData to display with RecyclerView. Must create an adapter to use with recyclerView. Also an item layout view must be created. Set the recyclerView adapter in the activity it will be displayed in, then set a layout manager for it (LinearLayout).

Removed implementation for database from the list view.